@@ -1,7 +1,17 @@
-/** @type {import("eslint").Linter.Config} */
-module.exports = {
-  extends: ["./base.js", "./extra.js"],
-  env: {
-    node: true,
+import globals from "globals";
+import baseConfig from "./base.js";
+
+/** @type {import('eslint').Flat.Config[]} */
+const config = [
+  ...baseConfig,
+  {
+    files: ["**/*.{js,ts,tsx}"],
+    languageOptions: {
+      globals: {
+        ...globals.node,
+      },
+    },
   },
-};
+];
+
+export default config;
